Fixing completely the int overflow bug in #202. Alignment was still not working properly for very large sequences: tested with sequence of PDB 2y9r (102 residues) against UniProt Q8WZ42 (34350 residues), a null pointer exception was happening The key in this patch is the modification in AbstractMatrixAligner: score has to be initialized to Integer.MIN_VALUE and not to Short.MIN_VALUE

+++ biojava3-alignment/src/main/java/demo/DemoAlignProteins.java
@@ -28,8 +28,8 @@ public class DemoAlignProteins {
 
 		GapPenalty penalty = new SimpleGapPenalty();
 
-		short gop = 8;
-		short extend = 1;
+		int gop = 8;
+		int extend = 1;
 		penalty.setOpenPenalty(gop);
 		penalty.setExtensionPenalty(extend);
